Updated Gazelle JSON definitions.

+++ tracker.go
@@ -231,8 +231,8 @@ func (t *GazelleTracker) GetStats() (*StatsEntry, error) {
 	// return StatsEntry
 	stats := &StatsEntry{
 		Tracker:       t.Name,
-		Up:            uint64(s.Response.Stats.Uploaded),
-		Down:          uint64(s.Response.Stats.Downloaded),
+		Up:            s.Response.Stats.Uploaded,
+		Down:          s.Response.Stats.Downloaded,
 		Ratio:         ratio,
 		Timestamp:     time.Now(),
 		TimestampUnix: time.Now().Unix(),
